run_alternate script

+++ root/scripts/start.bash
@@ -6,7 +6,11 @@ echo "To kill script, use the following command:"
 echo "kill -9 $processstartid"
 for (( ; ; )); do
 	let i++
-	bash /config/scripts/download.bash 2>&1 | tee "/config/logs/script_run_${i}_$(date +"%Y_%m_%d_%I_%M_%p").log" > /proc/1/fd/1 2>/proc/1/fd/2
+	if [ -f /config/run_alternate ]; then
+		bash /config/scripts/download_musicbrainz.bash 2>&1 | tee "/config/logs/script_run_${i}_$(date +"%Y_%m_%d_%I_%M_%p").log" > /proc/1/fd/1 2>/proc/1/fd/2
+	else
+		bash /config/scripts/download.bash 2>&1 | tee "/config/logs/script_run_${i}_$(date +"%Y_%m_%d_%I_%M_%p").log" > /proc/1/fd/1 2>/proc/1/fd/2
+	fi
 	if [ -f "/config/logs/log-cleanup" ]; then
 		rm "/config/logs/log-cleanup"
 	fi
